+++ src/linear_algebra/C/vectors.c
@@ -4,7 +4,7 @@
 #include <math.h>
 
 
-float norm(int n, float v[]) {
+double norm(int n, double v[]) {
     double sum = 0;
 
     #pragma omp parallel for reduction(+:sum)
@@ -12,32 +12,32 @@ float norm(int n, float v[]) {
         sum += v[i] * v[i];
     }
 
-    return (float) sqrt(sum);
+    return (double) sqrt(sum);
 }
 
 /* Vector Addition */
-void v_add(float u[], float v[], int n, float result[]) {    
+void v_add(double u[], double v[], int n, double result[]) {    
     for (int i = 0; i < n; i++) {
         result[i] = u[i] + v[i];
     }
 }
 
 /* Vector Subtraction */
-void v_sub(float u[], float v[], int n, float result[]) {
+void v_sub(double u[], double v[], int n, double result[]) {
     for (int i = 0; i < n; i++) {
         result[i] = u[i] - v[i];
     }
 }
 
 /* Vector Multiplication by Scalar (serial) */
-void v_mult_s(float v[], float c, int n, float result[]) {
+void v_mult_s(double v[], double c, int n, double result[]) {
     for (int i = 0; i < n; i++) {
         result[i] = v[i] * c;
     }
 }
 
 /* Vector Multiplication by Scalar (parallel) */
-void v_mult_p(float v[], float c, int n, float result[]) {
+void v_mult_p(double v[], double c, int n, double result[]) {
     int i;
     #pragma omp parallel for
     for (i = 0; i < n; i++) {
@@ -46,8 +46,8 @@ void v_mult_p(float v[], float c, int n, float result[]) {
 }
 
 /* Vector Dot Product */
-float dot(float u[], float v[], int n) {
-    float sum = 0;
+double dot(double u[], double v[], int n) {
+    double sum = 0;
     for (int i = 0; i < n; i++) {
         sum += u[i] * v[i];
     }
@@ -56,7 +56,7 @@ float dot(float u[], float v[], int n) {
 }
 
 /* Vector Cross Product */
-void cross(float u[], float v[], int n, float result[]) {
+void cross(double u[], double v[], int n, double result[]) {
     if (n != 3) {
         return;
     }
@@ -67,23 +67,23 @@ void cross(float u[], float v[], int n, float result[]) {
 }
 
 /* Triple Vector Product */
-void triple(float u[], float v[], float w[], int n, float result[]) {
+void triple(double u[], double v[], double w[], int n, double result[]) {
     if (n != 3) {
         return;
     }
     
-    float temp[3];
+    double temp[3];
     cross(v, w, n, temp);
     cross(u, temp, n, result);
 }
 
-void v_hadamard_s(float u[], float v[], int n, float result[]) {
+void v_hadamard_s(double u[], double v[], int n, double result[]) {
     for (int i = 0; i < n; i++) {
         result[i] = u[i] * v[i];
     }
 }
 
-void v_hadamard_p(float u[], float v[], int n, float result[]) {
+void v_hadamard_p(double u[], double v[], int n, double result[]) {
     int i;
     #pragma omp parallel for
     for (i = 0; i < n; i++) {
@@ -91,7 +91,7 @@ void v_hadamard_p(float u[], float v[], int n, float result[]) {
     }
 }
 
-void v_outer_s(float u[], float v[], int n, float result[n][n]) {
+void v_outer_s(double u[], double v[], int n, double result[n][n]) {
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
             result[i][j] = u[i] * v[j];
@@ -99,7 +99,7 @@ void v_outer_s(float u[], float v[], int n, float result[n][n]) {
     }
 }
 
-void v_outer_p(float u[], float v[], int n, float result[n][n]) {
+void v_outer_p(double u[], double v[], int n, double result[n][n]) {
     int i, j;
     #pragma omp parallel for collapse(2)
     for (i = 0; i < n; i++) {
