+++ src/root_finding/C/newton.c
@@ -1,7 +1,7 @@
 #include <math.h>
 #include "functions.h"
 
-double newton(double x0, double (*f)(double x), double (*f_prime)(double x), float tol, int iters) {
+double newton(double x0, double (*f)(double x), double (*f_prime)(double x), double tol, int iters) {
     double error = 10 * tol;
     int i = 0;
     while (error > tol && i < iters) {
